Extract ie-generate core
